requirements.txt
attrs==25.3.0
Automat==25.4.16
boto3==1.38.32
botocore==1.38.32
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
constantly==23.10.4
cryptography==45.0.3
cssselect==1.3.0
defusedxml==0.7.1
filelock==3.18.0
greenlet==3.2.3
h11==0.16.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itemadapter==0.11.0
itemloaders==1.3.2
jmespath==1.0.1
lxml==5.4.0
narwhals==1.41.1
numpy==2.2.6
packaging==25.0
pandas==2.3.0
parsel==1.10.0
plotly==6.1.2
Protego==0.4.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyDispatcher==2.0.7
pyOpenSSL==25.1.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
queuelib==1.8.0
requests==2.32.3
requests-file==2.1.0
s3transfer==0.13.0
Scrapy==2.13.1
service-identity==24.2.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.41
tldextract==5.3.0
Twisted==25.5.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
w3lib==2.3.1
websocket-client==1.8.0
wsproto==1.2.0
zope.interface==7.2
python-dotenv
python-dotenv
python-dotenv
